requirements.txt
streamlit>=1.35.0
plotly
joblib
google-auth
altair<5.0.0
protobuf==3.20.3
scikit-learn
pandas==2.2.2
numpy==1.26.4
google-cloud-bigquery==3.21.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
db-dtypes==1.2.0
lxml==5.2.1
urllib3==2.2.1
folium
streamlit-folium